requirements.txt
# Core
django==4.2.9
python-dotenv==1.0.1
str2bool==1.1

# UI 
django-admin-tabler==1.0.2

# Deployment
whitenoise==6.6.0
gunicorn==21.2.0
django-debug-toolbar==3.2.2
django-ckeditor
pillow
djangorestframework
djangorestframework-simplejwt
python-docx
django-import-export
django-rangefilter
lxml
openpyxl
psycopg2-binary
# mysqlclient
